Add overTime and attendance report

+++ HR_System_Backend/Model/OverTime.cs
@@ -14,7 +14,7 @@ namespace HR_System_Backend.Model
 
         public int OverTimeId { get; set; }
         public DateTime? OverTimeDate { get; set; }
-        public int? OverTimeHours { get; set; }
+        public double? OverTimeHours { get; set; }
         public double? OverHourPrice { get; set; }
         public double? OverTimePercentage { get; set; }
         public double? OverTimeTotal { get; set; }
